// src/app/models/masters/address.ts
export class Address {
    id: number;
    idAddressType: number;
    idHousingType: number;
    idCity: number;
    idMunicipality: number;
    idState: number;
    idCountry: number;
    addressType: string;
    housingType: string;
    city: string;
    municipality: string;
    state: string;
    country: string;
    avenue: string;
    street: string;
    edifice: string;
    floor: string;
    apartament: string;    
    references: string;
    postalCode: number;
    latitude: number;
    length: number;
}